Full update to Angular 6+, revamped project layout for ng6 workspace with the ng-canvas-guages library project and a demo project.
