--- .npmignore
@@ -1,2 +1,0 @@
-examples
-node_modules
