--- example/src/app/app.component.html
@@ -1,49 +1,0 @@
-<h1>
-  {{title}}
-</h1>
-
-<linear-gauge
-    width="150"
-    height="400"
-    units="°C"
-    title="Temperature"
-    min-value="-50"
-    max-value="50"
-    major-ticks="[-50,-40,-30,-20,-10,0,10,20,30,40,50]"
-    minor-ticks="5"
-    stroke-ticks="true"
-    ticks-width="15"
-    ticks-width-minor="7.5"
-    highlights='[ {"from": -50, "to": 0, "color": "rgba(0,0, 255, .3)"},
-    {"from": 0, "to": 50, "color": "rgba(255, 0, 0, .3)"} ]'
-    color-major-ticks="#ffe66a"
-    color-minor-ticks="#ffe66a"
-    color-title="#eee"
-    color-units="#ccc"
-    color-numbers="#eee"
-    color-plate="#2465c0"
-    color-plate-end="#327ac0"
-    border-shadow-width="0"
-    highlights='[
-            { "from": -50, "to": -30, "color": "rgba(0,255,0,.05)" },
-            { "from": -30, "to": -20, "color": "rgba(255,255,0,.25)" },
-            { "from": -20, "to": -10, "color": "rgba(255,30,0,.35)" },
-            { "from": -10, "to": 0, "color": "rgba(255,0,225,.45)" },
-            { "from": 0, "to": 50, "color": "rgba(0,0,255,.55)" }
-        ]'
-    borders="false"
-    border-radius="10"
-    needle-type="arrow"
-    needle-width="3"
-    animation-duration="250"
-    animation-rule="linear"
-    animated-value="true"
-    color-needle="#222"
-    color-needle-end=""
-    color-bar-progress="#327ac0"
-    color-bar="#f5f5f5"
-    bar-stroke="0"
-    bar-width="8"
-    bar-begin-circle="false"
-    [attr.value]="value"
-  ></linear-gauge>
