--- example/src/app/app.module.ts
@@ -1,23 +1,0 @@
-import {BrowserModule} from '@angular/platform-browser';
-import {NgModule} from '@angular/core';
-import {FormsModule} from '@angular/forms';
-import {HttpModule} from '@angular/http';
-
-import {AppComponent} from './app.component';
-import {GaugesModule} from 'ng-canvas-gauges';
-
-@NgModule({
-    declarations: [
-        AppComponent,
-    ],
-    imports: [
-        BrowserModule,
-        FormsModule,
-        HttpModule,
-        GaugesModule
-    ],
-    providers: [],
-    bootstrap: [AppComponent]
-})
-export class AppModule {
-}
