--- example/src/typings.d.ts
@@ -1,4 +1,0 @@
-/* SystemJS module definition */
-declare var module: {
-  id: string;
-};
