+++ package.json
@@ -1,48 +1,72 @@
 {
   "name": "ng-canvas-gauges",
-  "version": "2.0.3",
-  "description": "Angular 2 component wrapper for the canvas-gauges lib written by @Mikhus",
-  "main": "lib/index.js",
+  "version": "6.0.0",
+  "description": "Angular 6+ component wrapper for the canvas-gauges lib written by @Mikhus",
   "scripts": {
-    "watch": "tsc -p src -w",
-    "build": "rm -rf lib && tsc -p src"
+    "build:example": "ng build example",
+    "build:lib": "ng build ng-canvas-gauges",
+    "prebuild:lib": "npm run clean",
+    "postbuild:lib": "cp -f ./LICENSE  ./README.md ./dist/ng-canvas-gauges",
+    "build:pack": "cd dist/ng-canvas-gauges && npm pack",
+    "build:package": "npm run build:lib && npm run build:pack",
+    "postbuild:package": "mkdir -p lib && cp dist/ng-canvas-gauges/ng-canvas-gauges-*.tgz lib",
+    "clean": "rm -rf dist/ lib/",
+    "start": "ng serve"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/MeetmeLeave/ng-canvas-gauges.git"
   },
   "keywords": [
-    "Angular 2",
+    "Angular",
     "Canvas",
-    "Gauges"
+    "Gauges",
+    "Thermometer",
+    "Meter",
+    "Speedometer"
   ],
   "author": "vladimir.martynenko.work@gmail.com",
+  "contributors": [
+    {
+      "name": "Wayne Parrott",
+      "url": "http://wayneparrott.com"
+    }
+  ],
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/MeetmeLeave/ng-canvas-gauges/issues"
   },
   "homepage": "https://github.com/MeetmeLeave/ng-canvas-gauges",
-  "devDependencies": {
-    "@types/core-js": "0.9.36",
-    "typescript": "^2.4.0",
-    "typings": "^2.1.0"
-  },
   "dependencies": {
-    "@angular/animations": "^4.2.5",
-    "@angular/common": "^4.2.5",
-    "@angular/compiler": "^4.2.5",
-    "@angular/compiler-cli": "^4.2.5",
-    "@angular/core": "^4.2.5",
-    "@angular/forms": "^4.2.5",
-    "@angular/http": "^4.2.5",
-    "@angular/platform-browser": "^4.2.5",
-    "@angular/platform-browser-dynamic": "^4.2.5",
-    "@angular/platform-server": "^4.2.5",
-    "@angular/router": "^4.2.5",
-    "@types/canvas-gauges": "^2.1.0",
-    "canvas-gauges": "^2.1.4",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.2.0",
-    "zone.js": "^0.8.12"
+    "@angular/animations": "~7.2.0",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "~7.2.0",
+    "@angular/router": "~7.2.0",
+    "@types/canvas-gauges": "^2.1.1",
+    "canvas-gauges": "^2.1.5",
+    "core-js": "^2.5.4",
+    "rx-dom-html": "^7.0.3",
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.8.26"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.12.0",
+    "@angular-devkit/build-ng-packagr": "~0.12.0",
+    "@angular/cli": "~7.2.3",
+    "@angular/compiler-cli": "~7.2.0",
+    "@angular/language-service": "~7.2.0",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.5.0",
+    "ng-packagr": "^4.2.0",
+    "ts-node": "~7.0.0",
+    "tsickle": ">=0.34.0",
+    "tslib": "^1.9.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.2.2"
   }
 }
