+++ projects/ng-canvas-gauges/src/public_api.ts
@@ -1,0 +1,9 @@
+/*
+ * Public API Surface of ng-canvas-gauges
+ */
+export * from './lib/gauges.module';
+
+// components
+export * from './lib/linear-gauge';
+export * from './lib/radial-gauge';
+
