--- src/radial-gauge.ts
@@ -1,44 +1,0 @@
-/*!
- * The MIT License (MIT)
- *
- * Copyright (c) 2017 Vlad Martynenko <vladimir.martynenko.work@gmail.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-import {Component, NgZone, ElementRef} from '@angular/core';
-import {BaseGauge} from './base-gauge';
-import * as CanvasGauges from 'canvas-gauges';
-
-/**
- * Implements Radial Gauge from the original library
- */
-@Component({
-    selector: 'radial-gauge',
-    template: '<canvas #gauge></canvas>'
-})
-export class RadialGaugeComponent extends BaseGauge<CanvasGauges.RadialGauge,CanvasGauges.RadialGaugeOptions> {
-    constructor(el: ElementRef, zone: NgZone) {
-        super(el, zone);
-    }
-
-    ngOnInit() {
-        this.gauge = new CanvasGauges.RadialGauge(this.getOptions()).draw();
-    }
-}
