--- src/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true,
-        "skipLibCheck": true,
-        "target": "es5",
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "removeComments": true,
-        "sourceMap": true,
-        "outDir": "../lib",
-        "declaration": true
-    },
-    "types": [
-        "core-js"
-    ]
-}
