--- typings.json
@@ -1,7 +1,0 @@
-{
-    "globalDependencies": {
-        "core-js": "registry:dt/core-js#0.0.0+20160725163759",
-        "jasmine": "registry:dt/jasmine#2.2.0+20160621224255",
-        "node": "registry:dt/node#6.0.0+20160909174046"
-    }
-}
